chore: minor changes review#1
